Create room feature (#18)
* [HM] | feature-create-room-feature
-add feature to create-room

+++ src/main/java/io/github/hotelmanagement/model/room/RoomDTO.java
@@ -1,11 +1,7 @@
 package io.github.hotelmanagement.model.room;
 
 import io.github.hotelmanagement.model.reservation.ReservationDTO;
-import lombok.AllArgsConstructor;
 import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.springframework.lang.NonNull;
 
 import java.util.List;
 
@@ -14,8 +10,7 @@ public record RoomDTO(
         Long id,
         int bedAmount,
         int maxPeopleInside,
-        int pricePerNight,
-
+        double pricePerNight,
         List<ReservationDTO> reservationDTOS
 )
 {
